stream nPrint results to subsequent pipeline step(s) as in-memory buffers
The user is relieved of having to store nPrint result files on disk
(even temporarily). Should they want these results saved (for debugging
purposes) they may opt in to this.

Step- and lower-level timing information is also added to the pipeline
results.

Note: This is part of an nPrint-nprintML optimization effort.

See: #39